[CBR] implements all the APIs of Tag. (#371)
[CBR] implements all the APIs of Tag.

What this PR does / why we need it
The customer wants to have the possibility to update CBR Vault tags.
https://jira.tsi-dev.otc-service.com/browse/DM-3663
Which issue this PR fixes

fixes #370
Special notes for your reviewer
I cannot write the corresponding test
you can either implement it or wait for my supervisor to arrange an available OTC environment for me.

Reviewed-by: Anton Sidelnikov <None>
Reviewed-by: Artem Lifshits <None>
Reviewed-by: Aloento <None>
Reviewed-by: Artem Goncharov <None>